+++ support/compile.js
@@ -46,6 +46,7 @@ function parseRequires(js) {
     .replace(/require\('events'\)/g, "require('browser/events')")
     .replace(/require\('debug'\)/g, "require('browser/debug')")
     .replace(/require\('path'\)/g, "require('browser/path')")
+    .replace(/require\('diff'\)/g, "require('browser/diff')")
     .replace(/require\('tty'\)/g, "require('browser/tty')")
     .replace(/require\('fs'\)/g, "require('browser/fs')")
 }