+++ test/acceptance/timeout.js
@@ -14,7 +14,9 @@ describe('timeouts', function(){
 
   it('should allow overriding per-test', function(done){
     this.timeout(1000);
-    setTimeout(done, 300);
+    setTimeout(function(){
+      done();
+    }, 300);
   })
 })