+++ test/browser/opts.js
@@ -1,0 +1,5 @@
+describe('Options', function() {
+  it('should set timeout value', function() {
+    assert(this._test._timeout === 1500);
+  });
+})