+++ test/hook.timeout.js
@@ -1,0 +1,9 @@
+
+before(function(done){
+  this.timeout(100);
+  setTimeout(done, 50);
+})
+
+it('should work', function(done) {
+  done();
+});